Create "CreateProjectForm" component skeleton.
Built the "CreateProjectForm" component, conformed by a Modal dialog
which contains three label-input pairs for the user to specify the
project's data, plus two buttons, one to close the dialog and another
one to save the project.

Gave the component basic styling with CSS and Tailwind, and used the
"createPortal" method to link it to the "modal-root" div from the
"index.html" file, facilitating its styling.

Positioning of the modal is pending to be finished in the future.