Add interactive display for "CreateProjectForm" component.
Turned the buttons from the "MainDisplay" and "ProjectsSidebar" into
interactive so that on a user click they display the "CreateProjectForm"
replacing the "MainDisplay" one.

This is accomplished by using a "State" hook within the App component
(parent of both components where the buttons lie) triggered by the
"onClick" prop in the mentioned buttons.

The buttons from the Dialog in the "CreateProjectForm" component also
use that state to revert the effect, so that when the dialog is closed
or correctly submitted, the "MainDisplay" component takes his place back
from the "CreateProjectForm" one.

The styling in the "index.css" file was also modified to eliminate an
unwanted interaction by which the "modal-root" div was blocking the
buttons from components "ProjectsSidebar" and "MainDisplay", so that
they weren't able to be clicked.