Build project-data-saving functionality.
Changes to "CreateProjectForm":

- This component have been updated by adding three new refs that gather
data from the input and textarea elements.

- Also, now the "onSubmit" prop is not passed on to the main "App"
component to execute a function there. Instead, now it executes new
function "handleSubmit" within the "CreateProjectForm" component itself.

- Function "handleSubmit" creates an object "submittedProject" which
stores the corresponding ref data within keys "title", "description",
and "dueDate". This object is then passed down as the argument for new
function "onCreateProject", which is used as a Prop and passed down to
the main "App" component.

Changes to "App" component:

- Now "App" receives the "onCreateProject" prop from the
"CreateProjectForm" sub-component. This prop is used to trigger function
"handleSubmitProject" replacing the old "onSubmit" prop. Since the
"CreateProjectForm" prop is linked to the "handleSubmit" function from
the "CreateProjectForm" sub-component, which is triggered by an
"onSubmit" prop there, it essentially accomplishes executing the
"handleSubmitProject" project function when the user submits the modal.

- New Array "listOfProjects" is created outside of the "App" component.

- Function "handleSubmitProject" pushes the "submittedProject" Object
into the "listOfProjects" Array. This accomplishes an Array of objects
with the data of each created project. This array can now be passed onto
the "ProjectsSidebar" component in order to display de projects in the
sidebar.