Update "CreateProjectForm" to display the list of projects.
The "listOfProjects" Array is passed as a prop from the main "App"
component to the "CreateProjectForm" one. There it is transformed to add
an "id" key in each object. This id will be equal to the object's index
position within the array.

The "listOfProjects" transformed array is afterwards mapped within a "ul"
element to create a "li" element for each object in it. A button is also
nested within the "li" element. This button has the current project's
"title" key as text.

Some styling was added to the list of projects with TailWind.