Correct "CreateProjectForm" style and delete portal.
The "CreateProjectForm" component was not correctly positioned in the
display and the fundamental cause for this was that it was being
forwarded via portal to the "root-modal" div. Due to this issue, the
portal has been deleted and "CreateProjectForm" is now a div component
instead of a modal.

The same styling as the one in the main div of the "MainDisplay"
component has been added to "CreateProjectForm". This accomplishes that
the positioning of both components in the display is exactly the same.

Finally, the "input" and "label" styles have been stored within two
variables which are used in the corresponding elements, to avoid
repetition.