Build the Project Overview functionality.
Project Overview functionality means that when the user clicks one of
the projects listed within the "ProjectsSidebar" component, the main
display shows the information of this Project. In order to accomplish
that, the following steps were taken:

- Create new component "ProjectOverview" basic skeleton.

This component will show the data used by the user to create the project
(Title, Description and Due Date) and will have an input element to add
the tasks needed to accomplish the project. User's will be able to add
and delete tasks, as well as delete the whole Project (these features
are still pending to be implemented).

- New prop "onProjectSelection" is implemented within the Project's listed
buttons within the "ProjectsSidebar" component and passed onto the "App"
component, so there it can trigger a new function to update the display
and show the new "ProjectOverview" component.

- The App component has been reworked to properly accept the three
interactive components which share the main display.

Now the "useState" hook's default value is "undefined" instead of "false".
When the user clicks the buttons that open the "CreateProjectForm"
component, the useState variable changes to "null" instead of "true".
When the user selects a project of the list, new function
"handleProjectSelection" changes the value of "useState" to true.

A new variable "mainDisplay" has been created and is displayed as the
value for the main display. This value will be equal to the
"DefaultDisplay" component (renamed from "MainDisplay" for clarity
purposes) by default. An if ... else condition will change the value of
"mainDisplay" to component "CreateProjectForm" when the "useState"
variable is "null", and to component "ProjectOverview" when it is "true"
instead. This effectively accomplishes the transition between the
different components that need to be shown as the main display.