Build list of tasks functionality within "ProjectOverview".
A new useState hook is created within the "ProjectOverview" component in
order to handle the addition or deletion of tasks via the input element.

The useState manages variable "listOfTasks". This variable is an empty
array on its default state and adds or deletes tasks to and from it by
executing functions "createTask" or "deleteTask". This array is mapped
so that the list of tasks is always shown within the display. If the
array is empty no content is displayed, but when the user adds tasks
those will be displayed alongside the "clear" button, which will be the
trigger for the "deleteTask" function.

Also, a useRef hook is used and set to the input element, so that its
value is always available and can be used to add tasks to the array.