Display specific project data at "ProjectOverview".
The "ProjectsSidebar" component now passes an "id" prop to the projects
listed. This id is equivalent to the project's index within the
"listOfProjects" array, and can then be used within the "App" component.
When the user clicks a project, a new State is executed (with a default
value of "undefined") and given the value of the project's id by calling
it from the button's click event. This new State variable is called
"selectedProject" and is used to mark which of the project's within
the "listOfProjects" array must be passed onto the "ProjectOverview"
component via the "selectedProject" prop. Now the three main data
elements from the "ProjectOverview" component are filled with the data
from the "selectedProject" prop, which corresponds with the data for
the Project that the user clicked.