Lift "listOfTasks" state up to handle tasks for every project.
The "listOfTasks" state was not correctly used within the
"ProjectOverview" component because it needed to interact with multiple
functions/variables from the main "App" component. Thus the state has
been lifted to accomplish the same functionality.

The Ref from the input which allows to create the tasks within the
"ProjectOverview" component is now forwarded from there to "App".

Functions "createTask" and "deleteTask" have been moved to the "App"
component and rework to fit.

"createTask" now updates the state depending on two conditions:

--> First one is if the "listOfProjects" current project Object has
already a "tasks" key created and if that key holds an array with a
length superior to 0. If that is the case, this means that the project
has already tasks created, and then the value of the state is updated
by copying not the previous value of the state but the value of the
mentioned "tasks" key, and then is added the current task by
using the imported Ref. That way the already created tasks are not lost
if the user changes between projects and the previous "listOfTasks" value
comes from another project.

-->Second is when the "listOfProjects" current project does not have yet
any tasks. In this case, the "listOfTasks" is pushed the new task and
then is used to create the "tasks" key within the Object corresponding
to the current project within variable "listOfProjects".

"deleteTask" works exactly as previously within the "ProjectOverview"
component.

The "listOfTasks" state is also executed from function
"handleProjectSelection" so that the array's values can be cleared when
the user changes the project. Those values are already saved at the
"listOfProjects" variable and that way we ensure there is not an unwanted
tasks copy from one project to another.

Finally, the Object from array "listOfProjects" that corresponds to the
current project is passed to the "ProjectOverview" component through
prop "SelectedProject" so that it can be used there to display the
information of the project and map the "tasks" key. Since every time
the state is updated because of project information changes, this prop
information will always be updated and the "ProjectOverview" component
always displays the correct information.