Build project deletion functionality.
Now projects can be deleted permanently from the "ProjectOverview"
display by using the "delete" button as a catalyst. This button is
given a hardcoded id and an onClick prop so that when the user clicks it,
function "handleCloseProjectDisplay" (previously knows as
"handleCloseCreateProject") is triggered. This function changes the
"display" state (previously known as "createProject") to "undefined" so
that the "DefaultDisplay" is shown instead of the "ProjectOverview"
component. Now this function also checks if the button click has the
"deleteProject" id, and if it is so it deletes the current project data
from the "listOfProjects" array, essentialy erasing it permanently.

Function name changes were made to reflect better their functionality.