Fix backgroundColor style for the tasks div in "ProjectOverview".
Created variable "tasksDivStyling" which now holds the style values for
the tasks div and used it as the new value for the "className" prop.

This variable is updated conditionally so that if the "listOfTasks" array
has tasks, the div will have a grey background colour, and otherwise it
will be white.