Fix "createTask" function bug.
"createTask" function was not allowing to create more than 1 task per
project dut to the added if statement in the last update and how
checking the input value to be true was interacting with setting its
value to null after the task was created. Now the value is set to ""
instead and this allows for a successful check, which alows tasks to be
created as expected.