Final styling tweaks.
Updated the color for the "ProjectOverview" buttons text.

Deleted outline for "CreateProjectForm" inputs when focused, and updated
the color of their bottom border in these conditions to a darker one.

Gave a new background and text color for the project list items created
within the "ProjectsSidebar" component when focused. In order to do that
I had to create a new const "liSelectedStyle" with the styles for the
focused project list item. These styles are set conditionally only to
the list item of the project which is selected by etsting if the
"projectIndex" mapping key and the "isSelected" prop are equal. The
"isSelected" prop is new and derives from the "App" component, being
equal to the "selectedProject" variable value (id of the project). Now
"selected" project is reset to its default "undefined" value when the
user deletes a project or opens the "CreateProjectForm" display, so that
the list item with the focus style goes back to its default styling.